backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiomysql==0.2.0
aiosignal==1.3.2
aiosqlite==0.21.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
Babel==2.10.3
backoff==2.2.1
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
cbor2==5.6.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.8
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
docutils==0.20.1
durationpy==0.10
et_xmlfile==2.0.0
fastapi==0.115.9
python-multipart==0.0.6
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
freezegun==1.2.1
frozenlist==1.6.0
fsspec==2025.5.0
geoip2==2.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
griffe==1.7.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.6
importlib_metadata==8.6.1
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
libsass==0.22.0
lief==0.17.0
lightgbm==4.6.0
llama-cloud==0.1.22
llama-cloud-services==0.6.23
llama-index-agent-openai==0.4.8
llama-index-cli==0.4.1
llama-index-core==0.12.37
llama-index-embeddings-huggingface==0.5.4
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.42
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.23
lxml==5.2.1
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
maxminddb==2.6.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
num2words==0.5.13
numpy==2.2.6
oauthlib==3.2.2
ofxparse==0.21
onnxruntime==1.22.0
openai==1.81.0
openpyxl==3.1.2
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.3.7
polib==1.1.1
posthog==4.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.8
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMySQL==1.1.2
pyOpenSSL==24.1.0
pyparsing==3.2.3
pypdf==5.5.0
PyPDF2==2.12.1
PyPika==0.48.9
pypiwin32==223
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.1.0
python-stdnum==1.19
pytz==2025.2
pyusb==1.2.1
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.1.0
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rjsmin==1.2.0
rl-renderPM==4.0.3
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
starlette==0.45.3
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.2
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.34.2
vobject==0.9.6.1
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
yara-python==4.5.4
yarl==1.20.0
yt-dlp==2025.10.14
zeep==4.2.1
zipp==3.21.0
zstandard==0.23.0
